import(tasks): delete deprecated and update hardhat config
