+++ contracts/dollar/tasks/library/incompatible-interface/manager.ts
