+++ contracts/dollar/tasks/library/incompatible-interface/priceReset.ts
@@ -5,7 +5,7 @@ import { task, types } from "hardhat/config";
 import { BondingV2 } from "../artifacts/types/BondingV2";
 import { IMetaPool } from "../artifacts/types/IMetaPool";
 import { UbiquityAlgorithmicDollarManager } from "../artifacts/types/UbiquityAlgorithmicDollarManager";
-import pressAnyKey from "../utils/flow";
+import pressAnyKey from "../../utils/flow";
 
 import { DEPLOYMENT_OVERRIDES, FORKING_CHAIN_NAME } from "./constants"
 import { A_PRECISION, get_burn_lp_amount } from "./utils"
@@ -35,8 +35,8 @@ task(
     ) => {
       console.log("started....");
 
-      // All the deployments in hardhat-deploy are stored in deployments directory. 
-      // There might be some cases we have to override them or use already deployed ones. 
+      // All the deployments in hardhat-deploy are stored in deployments directory.
+      // There might be some cases we have to override them or use already deployed ones.
       // In this case, they will be different per chainId. so that would be awesome to have them per chain.
 
       const OVERRIDES_PARAMS = DEPLOYMENT_OVERRIDES[network.name];
