+++ contracts/dollar/tasks/library/incompatible-interface/revoke.ts
