+++ contracts/dollar/tasks/library/investor-emissions.ts
@@ -2,7 +2,7 @@ import "@nomiclabs/hardhat-waffle";
 import { Wallet } from "ethers";
 import "hardhat-deploy";
 import { ActionType } from "hardhat/types/runtime";
-import { impersonate as action } from "./investor-emissions/impersonate";
+import { impersonate } from "./investor-emissions/impersonate";
 
 let distributor = "0x445115D7c301E6cC3B5A21cE86ffCd8Df6EaAad9";
 
@@ -16,15 +16,13 @@ export const addressBook = {
   receiver: "0xefC0e701A824943b469a694aC564Aa1efF7Ab7dd", // Investor
 };
 
-module.exports = {
-  description: "Distributes investor emissions",
-  params: {
-    receiver: "Account thats receiving the tokens",
-  },
-  optionalParams: {
-    token: ["Ubiquity governance token address", addressBook.token],
-    sender: ["Account thats distributing the tokens", addressBook.sender],
-    // receiver: ["Account thats receiving the tokens", addressBook.receiver],
-  },
-  action: (): ActionType<any> => action,
+export const description = "Distributes investor emissions";
+export const params = {
+  receiver: "Account thats receiving the tokens",
 };
+export const optionalParams = {
+  token: ["Ubiquity governance token address", addressBook.token],
+  sender: ["Account thats distributing the tokens", addressBook.sender],
+  // receiver: ["Account thats receiving the tokens", addressBook.receiver],
+};
+export const action = (): ActionType<any> => impersonate;
